Rename the linkin package to linkerd
